requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
PyHamcrest==1.9.0
requests==2.19.1
selenium==3.12.0
six==1.11.0
urllib3==1.23